Extract LFPFileReader tests into tests/unit/test_lfp_reader.py
